Move C++ `Config` to its own header
Also bind APIs which were added since this was created. Also copy over
some `#define` guards for the C methods into the C++ header as well. And
finally update tests for what's added.
